refactor dpm wave ai routes
